Add testing for Manager. Redo and undo buttons. Save state.
This is combo PR!

I didn't have a lot of unit testing around the manager
class. This will add testing to public methods.

It also adds redo and undo functionality.

Redo will only work if there is state that is redo-able.
The same is true for undo functionality. Also, if you
press the undo button, then another state button that
is neither undo or redo, all existing redo will be cleared
from state.

It also stores stats on the selectedGame and stores
allGames in UserDefaults. The app can be closed and
re-opened with saved state from the last close.
There is a test for this, but it does not work.
It passes this requirement in smoke testing.